Migrated seed type from int to long (API breaking change)
Resolves #207

+++ instancio-core/src/main/java/org/instancio/Random.java
@@ -28,7 +28,7 @@ public interface Random {
      *
      * @return seed value
      */
-    int getSeed();
+    long getSeed();
 
     /**
      * Returns a random boolean with a {@code 0.5} probability.
